Adicionado a função 'Atualizar Lista' referente a atuaslização da lista de professores.
