+++ aplicacoes/afastamento/src/main/java/ufc/quixada/npi/afastamento/jobs/AfastamentoScheduler.java
@@ -2,7 +2,6 @@ package ufc.quixada.npi.afastamento.jobs;
 
 import java.util.Calendar;
 import java.util.Date;
-import java.util.List;
 
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -14,7 +13,6 @@ import org.springframework.scheduling.annotation.Scheduled;
 
 import ufc.quixada.npi.afastamento.model.Afastamento;
 import ufc.quixada.npi.afastamento.model.Periodo;
-import ufc.quixada.npi.afastamento.model.Professor;
 import ufc.quixada.npi.afastamento.model.Ranking;
 import ufc.quixada.npi.afastamento.model.StatusPeriodo;
 import ufc.quixada.npi.afastamento.model.StatusReserva;
@@ -22,12 +20,8 @@ import ufc.quixada.npi.afastamento.model.StatusTupla;
 import ufc.quixada.npi.afastamento.model.TuplaRanking;
 import ufc.quixada.npi.afastamento.service.AfastamentoService;
 import ufc.quixada.npi.afastamento.service.PeriodoService;
-import ufc.quixada.npi.afastamento.service.ProfessorService;
 import ufc.quixada.npi.afastamento.service.RankingService;
 import ufc.quixada.npi.afastamento.service.ReservaService;
-import br.ufc.quixada.npi.ldap.model.Constants;
-import br.ufc.quixada.npi.ldap.model.Usuario;
-import br.ufc.quixada.npi.ldap.service.UsuarioService;
 
 @Named
 @Configurable
@@ -42,12 +36,6 @@ public class AfastamentoScheduler {
 	
 	@Inject
 	private ReservaService reservaService;
-	
-	@Inject
-	private UsuarioService usuarioService;
-	
-	@Inject
-	private ProfessorService professorService;
 	
 	@Inject
 	private AfastamentoService afastamentoService;
@@ -89,32 +77,10 @@ public class AfastamentoScheduler {
 				}
 			}
 		}
-		adicionaNovosProfessores();
+		
 		
 	}
 	
-	private void adicionaNovosProfessores() {
-		List<Usuario> usuarios = usuarioService.getByAffiliation(Constants.AFFILIATION_DOCENTE);
-		for(Usuario usuario : usuarios) {
-			Professor professor = professorService.getByCpf(usuario.getCpf());
-			if(professor == null) {
-				professor = new Professor();
-				professor.setCpf(usuario.getCpf());
-				professorService.save(professor);
-			}
-		}
-		atualizaVagas();
-		
-	}
 	
-	private void atualizaVagas() {
-		Periodo periodoAtual = periodoService.getPeriodoPosterior(periodoService.getPeriodoPosterior(periodoService.getPeriodoAtual()));
-		List<Periodo> periodos = periodoService.getPeriodosPosteriores(periodoAtual);
-		int vagas = (int) (professorService.findAtivos().size() * 0.15);
-		for(Periodo periodo : periodos) {
-			periodo.setVagas(vagas);
-			periodoService.update(periodo);
-		}
-	}
 
 }
