+++ aplicacoes/afastamento/src/main/java/ufc/quixada/npi/afastamento/util/Constants.java
@@ -41,6 +41,7 @@ public class Constants {
 	
 	public static final String REDIRECT_PAGINA_LISTAR_PERIODOS = "redirect:/administracao/periodo";
 	
+	public static final String REDIRECT_PAGINA_LISTAR_PROFESSORES = "redirect:/administracao/professores";
 	
 	/** Mensagens */
 	
@@ -68,5 +69,13 @@ public class Constants {
 	
 	public static final String MSG_ERRO_ATUALIZAR_PERIODO = "Ocorreu um erro na atualização do período.";
 	
+	/** Afiliações */
+
+	public static final String AFFILIATION_DISCENTE = "DISCENTE";
 	
+	public static final String AFFILIATION_DOCENTE = "DOCENTE";
+	
+	public static final String AFFILIATION_ADMIN_SIAF = "ADMIN-SIAF";
+	
+	public static final String AFFILIATION_STA = "STA";
 }
